requirements.txt
nltk
sklearn
requests
python-telegram-bot
pymorphy2
pymorphy2-dicts-ru
logging